renaming searchResultsPage class
