+++ todo-spec.js
@@ -1,5 +1,5 @@
 let aliexpressHomePage = require('./pages/aliexpressHomePage')
-let searchResultsPage = require('./pages/searchResultspage')
+let searchResultsPage = require('./pages/searchResultsPage')
 
 describe('searching for products in Aliexpress', function() {
   it('search for iphone product and check the 2nd page', async function() {
